requirements.txt
selenium==4.36.0
pytest 
requests
python-dotenv